feature/dep316: Fixed engagement status logic for engagement searches in API, removed unneeded revalidator, fixed metadata and engagement status parsing in landing loader, organized EngagementSearch components.
