+++ web/src/components/landing/index.tsx
@@ -1,10 +1,7 @@
 import React, { createContext, useEffect, useMemo, useState } from 'react';
-import { LandingData } from './types';
+import { LandingData, SortOrder } from './types';
 import { defaultLandingData, defaultSearchFilters } from './constants';
-import { useLoaderData, useRevalidator, useSearchParams } from 'react-router';
-import { ThemeProvider } from '@mui/material';
-import FilterDrawer from './FilterDrawer';
-import { DarkTheme } from 'styles/Theme';
+import { useLoaderData, useSearchParams } from 'react-router';
 import { LandingHero } from './LandingHero';
 import { TenantState } from 'reduxSlices/tenantSlice';
 import { useAppSelector } from 'hooks';
@@ -21,14 +18,27 @@ export const Landing = () => {
     const tenant: TenantState = useAppSelector((state) => state.tenant);
     const [searchParams, setSearchParams] = useSearchParams(); // Pass filter updates to loader via search params
     const { engagements, allMetaFilters } = useLoaderData();
-    const revalidator = useRevalidator();
 
     useEffect(() => {
-        revalidator.revalidate();
+        if (searchParams.size === 0) {
+            // Set default filters if none are found
+            clearFilters();
+        }
     }, [searchParams]);
 
     const clearFilters = () => {
-        const newSearchParams = updateSearchParams(defaultSearchFilters, new URLSearchParams());
+        const sortOrder = searchParams.get('sort_order') as SortOrder | undefined;
+        const dsf = defaultSearchFilters;
+        const newSearchParams = updateSearchParams(
+            {
+                // Retain search and sort, just remove filters
+                ...dsf,
+                search_text: searchParams.get('search_text') ?? dsf.search_text,
+                sort_key: searchParams.get('sort_key') ?? dsf.sort_key,
+                sort_order: sortOrder ?? dsf.sort_order,
+            },
+            new URLSearchParams(),
+        );
         setSearchParams(newSearchParams);
     };
 
@@ -59,9 +69,6 @@ export const Landing = () => {
 
     return (
         <LandingDataContext.Provider value={landingData}>
-            <ThemeProvider theme={DarkTheme}>
-                <FilterDrawer />
-            </ThemeProvider>
             <LandingHero />
             <LandingIntro />
             <EngagementSearch />
